% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionarytools-package.R
\docType{package}
\name{dictionarytools-package}
\alias{dictionarytools}
\alias{dictionarytools-package}
\title{dictionarytools: Tools for Text Analysis Dictionaries}
\description{
Tools for working with text analysis dictionaries.
}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@lse.ac.uk} [copyright holder]

Authors:
\itemize{
  \item Stefan Müller \email{mullers@tcd.ie}
}

}
\keyword{internal}
